+++ src/main/java/com/epeters/raytrace/Ray.java
@@ -8,11 +8,11 @@ import static com.epeters.raytrace.Vector.ORIGIN;
  */
 public record Ray(Vector origin, Vector direction) {
 
-    public Ray(float x, float y, float z) {
+    public Ray(double x, double y, double z) {
         this(ORIGIN, new Vector(x, y, z));
     }
 
-    public Vector at(float t) {
+    public Vector at(double t) {
         return origin.plus(direction.mul(t));
     }
 }
